refactor code

+++ pssetools/snp.py
@@ -1,14 +1,14 @@
+from __future__ import print_function
 import os
 from . import psspy
-from . import argument_parser
+from . import pss_activity
 
 _i = psspy.getdefaultint()
 _f = psspy.getdefaultreal()
 _s = psspy.getdefaultchar()
 
+@pss_activity
 def run(sav, snp, dyr, cc, ct, **kwargs):
-    psspy.case(sav)
-    
     # genero el primer dyr
     if len(dyr) == 0:
         raise ValueError("No hay *.dyr a cargar")
@@ -20,8 +20,8 @@ def run(sav, snp, dyr, cc, ct, **kwargs):
     # genero los dyr restantes
     tmp_cc = cc.replace(".flx", "_tmp.flx")
     tmp_ct = ct.replace(".flx", "_tmp.flx")
-    for dyr in dyr[1:]:
-        psspy.dyre_add([_i,_i,_i,_i], dyr, tmp_cc, tmp_ct)
+    for dyr_file in dyr[1:]:
+        psspy.dyre_add([_i,_i,_i,_i], dyr_file, tmp_cc, tmp_ct)
         
         with open(tmp_cc, "r") as conec: tmp_cc_lines = conec.readlines()
         with open(tmp_ct, "r") as conet: tmp_ct_lines = conet.readlines()
@@ -36,8 +36,8 @@ def run(sav, snp, dyr, cc, ct, **kwargs):
                 ct_lines.insert(-6, line)
                     
         # limpia los archivos temporales    
-        os.remove(tmp_cc)
-        os.remove(tmp_ct)
+        if os.path.exists(tmp_cc): os.remove(tmp_cc)
+        if os.path.exists(tmp_ct): os.remove(tmp_ct)
         
     with open(cc, "w") as conec: conec.writelines(cc_lines)
     with open(ct, "w") as conet: conet.writelines(ct_lines)
@@ -46,16 +46,4 @@ def run(sav, snp, dyr, cc, ct, **kwargs):
     # todo estas opciones deberian estar en otro lado
     psspy.dynamics_solution_param_2([_i,_i,_i,_i,_i,_i,_i,_i],[_f,_f, 0.002,_f,_f,_f,_f,_f])
     ierr = psspy.snap(sfile=snp)
-    assert ierr == 0
-        
-
-if __name__ == "__main__":
-    args_specs = {
-        "sav": {"type": str},
-        "snp": {"type": str}, 
-        "cc":  {"type": str}, 
-        "ct":  {"type": str}, 
-        "dyr": {"nargs": "*", "type": str}
-    }
-    args = argument_parser(args_specs)
-    run(**args)
+    return ierr
